Consolidate Gaze into module. Add Animation.
Extract functionality previously on MainWindow as it is better suited for
its own project. This leaves the MainWindow clean and decouples selection
as services, currently implemented for gaze.

Refactor as needed and create the service which may use the previously
created DataContainers and Filters (or not -- but they are there if
necessary).

Add filling-arc animation. As the selection service was tightly related to
it, the animation was included in this commit. Animation is left as a
resource (consider future refactoring to extract it to an infrastructure
dictionary) linked to a property; therefore other animations can be
created and used (i.e. user can select Animation at runtime).

Side-effects: As before, Suggestion Board is the guinea pig. Reason: It is
the smallest of boards and easiest to fix. (I might as well create a dummy
board for testing purposes, though.)