Improve animation handling by using the right VMs
Fixes an error where the DataContext for the wrong control was used.
Also sorts the logic of StateMachine processing to process all points,
including those which do not fall on a (Selectable)Control. Finally, it
uses a Dictionary to store references to the ViewModels instead of
retrieving them each time through the Dispatcher (remember to keep the use
of the Dispatcher at a minimum, as in microcontroller interruptions).