Add a SuggestionSource and changes for it to work
The suggestion source is Presage. It runs as a WCF service and therefore
there is no dependency on a library. It is necessary to install and run
Presage. More info on http://presage.sourceforge.net/

The required changes include creating and using a command specific for
suggestion selection. This way the entire process of clearing the
half-written word and adding space after the word can be easily
introduced.