Add Multikeyboard w/disambiguation & autocomplete
Includes an adjustment necessary on the suggestion service infrastructure
(namely ISuggestionService). Other changes to infrastructure include
addition of commands, a spanish dictionary sorted by frequency by the Real
Academia Espanola, events and constants.

Other projects affected are the SuggestionBoard which has to grab the
suggestions when necessary and SingleKeyBoard. The changes to
SingleKeyBoard are for testing purposes and should be discarded whenever
necessary.